api: refactor huskyCI routes and util package
