+++ api/routes/healthcheck.go
@@ -1,0 +1,12 @@
+package routes
+
+import (
+	"net/http"
+
+	"github.com/labstack/echo"
+)
+
+// HealthCheck is the heath check function.
+func HealthCheck(c echo.Context) error {
+	return c.String(http.StatusOK, "WORKING\n")
+}
