+++ api/util/util.go
@@ -2,14 +2,8 @@ package util
 
 import (
 	"bufio"
-	"crypto/tls"
-	"crypto/x509"
-	"io/ioutil"
-	"net/http"
 	"os"
 	"strings"
-
-	"github.com/globocom/huskyCI/api/log"
 )
 
 const (
@@ -19,44 +13,18 @@ const (
 	KeyFile = "api/api-tls-key.pem"
 )
 
-// NewClient returns an http client.
-func NewClient(httpsEnable bool) (*http.Client, error) {
-
-	if httpsEnable {
-		// Tries to find system's certificate pool
-		caCertPool, _ := x509.SystemCertPool() // #nosec - SystemCertPool tries to get local cert pool, if it fails, a new cer pool is created
-		if caCertPool == nil {
-			caCertPool = x509.NewCertPool()
+// RequestResponse returns a map string interface to be used in echo JSON requests response.
+func RequestResponse(success bool, errMsg string) map[string]interface{} {
+	if errMsg == "" {
+		return map[string]interface{}{
+			"success": success,
+			"error":   "none",
 		}
-
-		cert, err := ioutil.ReadFile(CertFile)
-		if err != nil {
-			log.Error("NewClientTLS", "UTIL", 4001, err)
-		}
-		if ok := caCertPool.AppendCertsFromPEM(cert); !ok {
-			log.Error("NewClientTLS", "UTIL", 4002, err)
-		}
-
-		tlsConfig := &tls.Config{
-			RootCAs: caCertPool,
-		}
-		tlsConfig.BuildNameToCertificate()
-		client := &http.Client{
-			Transport: &http.Transport{
-				TLSClientConfig: &tls.Config{
-					MinVersion:               tls.VersionTLS11,
-					MaxVersion:               tls.VersionTLS12,
-					PreferServerCipherSuites: true,
-					InsecureSkipVerify:       false,
-					RootCAs:                  caCertPool,
-				},
-			},
-		}
-		return client, nil
 	}
-
-	client := &http.Client{}
-	return client, nil
+	return map[string]interface{}{
+		"success": success,
+		"error":   errMsg,
+	}
 }
 
 // HandleCmd will extract %GIT_REPO% and %GIT_BRANCH% from cmd and replace it with the proper repository URL.
